Move Recipes into its own namespace
It is a lazyload block rather than a partial.
