+++ config/routes.yaml
@@ -45,10 +45,10 @@ programme_recommendations:
     defaults: { _controller: App\Controller\RecommendationsController, extension: '' }
     requirements: { pid: '[0-9b-df-hj-np-tv-z]{8,15}', extension: '$|\.2013inc' }
 
-partial_programme_recipes:
-    path: /programmes/{pid}/recipes.ameninc
-    defaults: { _controller: App\Controller\Partial\RecipesController }
-    requirements: { pid: '[0-9b-df-hj-np-tv-z]{8,15}' }
+programme_recipes:
+    path: /programmes/{pid}/recipes{extension}
+    defaults: { _controller: App\Controller\RecipesController, extension: '' }
+    requirements: { pid: '[0-9b-df-hj-np-tv-z]{8,15}', extension: '$|\.ameninc' }
 
 # Schedules
