+++ templates/find_by_pid/tlec.html.twig
@@ -74,7 +74,7 @@
                 <ul class="grid">
                     {%- if programme.getOption('recipes_enabled') -%}
                         <li class="grid__item 1/2@gel3 1/3@gel3b 1/4@gel5 gel-brevier-bold lazy-module box-link"
-                             data-lazyload-inc="{{ path('partial_programme_recipes', {'pid': programme.getPid()}) }}"
+                             data-lazyload-inc="{{ path('programme_recipes', {'pid': programme.getPid(), 'extension': '.ameninc'}) }}"
                              data-lazyload-threshold="320"
                              data-lazyload-delay="true"
                              data-object-type="recipe">
